rename action containers and cleanup imports
