+++ src/app/store/actions.ts
@@ -1,5 +1,5 @@
 export * from './actions/app.actions';
-export * from './actions/node.action';
-export * from './actions/snackbar.action';
-export * from './actions/router.action';
-export * from './actions/viewer.action';
+export * from './actions/node.actions';
+export * from './actions/snackbar.actions';
+export * from './actions/router.actions';
+export * from './actions/viewer.actions';
