+++ src/app/store/actions/app.actions.ts
@@ -4,7 +4,6 @@ export const SET_APP_NAME = 'SET_APP_NAME';
 export const SET_HEADER_COLOR = 'SET_HEADER_COLOR';
 export const SET_LOGO_PATH = 'SET_LOGO_PATH';
 
-
 export class SetAppNameAction implements Action {
     readonly type = SET_APP_NAME;
     constructor(public payload: string) {}
