+++ src/app/store/actions/viewer.actions.ts
